refactor: 중복된 클거스 MemberFixture 제
